refactor : 정산(각 회식) 참여 정책에 맞도록 수정(gathering -> bill, 각 회식 별로 참여 여부 결정)
