+++ src/main/kotlin/com/server/dpmcore/bill/bill/application/BillCommandService.kt
@@ -5,6 +5,7 @@ import com.server.dpmcore.bill.bill.domain.model.Bill
 import com.server.dpmcore.bill.bill.domain.model.BillId
 import com.server.dpmcore.bill.bill.domain.port.outbound.BillPersistencePort
 import com.server.dpmcore.bill.bill.presentation.dto.request.CreateBillRequest
+import com.server.dpmcore.bill.bill.presentation.dto.request.UpdateGatheringJoinsRequest
 import com.server.dpmcore.bill.billAccount.application.BillAccountQueryService
 import com.server.dpmcore.bill.billAccount.domain.model.BillAccountId
 import com.server.dpmcore.bill.exception.BillException
@@ -101,4 +102,9 @@ class BillCommandService(
         billId: BillId,
         memberId: MemberId,
     ) = gatheringCommandUseCase.submitBillParticipationConfirmEachGathering(billId, memberId)
+
+    fun markAsJoinedEachGathering(
+        request: UpdateGatheringJoinsRequest,
+        memberId: MemberId,
+    ) = gatheringCommandUseCase.markAsJoinedEachGatheringMember(request, memberId)
 }
