+++ src/main/kotlin/com/server/dpmcore/gathering/gathering/presentation/controller/GatheringApi.kt
@@ -2,9 +2,7 @@ package com.server.dpmcore.gathering.gathering.presentation.controller
 
 import com.server.dpmcore.common.exception.CustomResponse
 import com.server.dpmcore.gathering.gathering.domain.model.GatheringId
-import com.server.dpmcore.gathering.gathering.presentation.request.UpdateGatheringJoinsRequest
 import com.server.dpmcore.gathering.gathering.presentation.response.GatheringMemberJoinListResponse
-import com.server.dpmcore.member.member.domain.model.MemberId
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.media.Content
 import io.swagger.v3.oas.annotations.media.ExampleObject
@@ -16,25 +14,6 @@ import org.springframework.http.MediaType.APPLICATION_JSON_VALUE
 
 @Tag(name = "Gathering", description = "회식 API")
 interface GatheringApi {
-    @ApiResponse(
-        responseCode = "204",
-        description = "각 회식 참여 추가",
-        content = [
-            Content(
-                mediaType = APPLICATION_JSON_VALUE,
-                schema = Schema(implementation = CustomResponse::class),
-            ),
-        ],
-    )
-    @Operation(
-        summary = "각 회식 참여 추가",
-        description = "여러 회식의 참여 여부를 표시합니다.",
-    )
-    fun markAsJoined(
-        request: UpdateGatheringJoinsRequest,
-        memberId: MemberId,
-    ): CustomResponse<Void>
-
     @ApiResponse(
         responseCode = "200",
         description = "회식별 멤버 참여 여부 조회 성공",
